requirements.txt
fastapi
uvicorn[standard]
msal
httpx
python-dotenv
